OrderConfirmation refactoring
