+++ src/test/java/pageComponents/OrderSummaryPageComponent.java
@@ -1,0 +1,4 @@
+package pageComponents;
+
+public class OrderSummaryPageComponent {
+}
